requirement.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
pydantic
docx2txt
python-docx
PyMuPDF
zipstream-new
loguru